chore: Renaming utils and removing deprecated portions of test setup
