+++ packages/babel-plugin-ember-test-metadata/__tests__/get-normalized-file-path-test.js
@@ -1,4 +1,4 @@
-const { getNormalizedFilePath } = require('../utils');
+const { getNormalizedFilePath } = require('../get-normalized-file-path');
 
 describe('getNormalizedFilePath', () => {
   describe('classic', () => {
