+++ packages/babel-plugin-ember-test-metadata/get-normalized-file-path.js
